chore(schema): separated schemas
Separated schemas in each file
